Made Ray2 class (temp name):
Ray takes two points on a line (a hexagon side) and finds their midpoint. This is the starting position.

It also takes an int, which corresponds to a direction in its Direction enum.

Depending on the direction, its head (the front of the ray,the bit that moves) keeps moving

rayCheck method in HexagonGrid runs the contains() method on all hexagons, with the ray's head being taken as input

If the ray head is no konger inside of any hexagon, it stops moving (boolean isInside switched to false)