Multiple Atom interaction now works.
Important to note that there's a pattern to atom placements / ray reflection:
   - if there are more than 2 atoms the ray will ALWAYS reflect
   - if there are 2 atoms the ray will Deflect ONLY IF those two atoms are also neighbouring each other

A counter is made in Hexagon which counts how many atoms are neighbouring it.

When a ray enters a neighbour Hexagon setDirection() is called. If neighbCount is 2, multAtomDeflect() is called, which calls getSurroundingHexagons(). This method creates a list of all the hexagons around the ray hex. if there are two atom hexagons directly beside each other, it returns the first's index. Otherwise returns 10.

Depending on the return value and starting direction, the ray's direction is changed.